+++ packages/core/src/utils/converter.ts
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0.
 /* eslint-disable no-bitwise */
 import { Base64 } from "../encoding/base64";
+import { Base64Url } from "../encoding/base64Url";
 import { HexHelper } from "../helpers/hexHelper";
 
 /**
@@ -230,6 +231,24 @@ export class Converter {
 	 */
 	public static base64ToBytes(base64: string): Uint8Array {
 		return Base64.decode(base64);
+	}
+
+	/**
+	 * Convert bytes to base64 url string.
+	 * @param bytes The bytes to convert.
+	 * @returns A base64 url string of the bytes.
+	 */
+	public static bytesToBase64Url(bytes: Uint8Array): string {
+		return Base64Url.encode(bytes);
+	}
+
+	/**
+	 * Convert a base64 url string to bytes.
+	 * @param base64Url The base64 url string.
+	 * @returns The bytes.
+	 */
+	public static base64UrlToBytes(base64Url: string): Uint8Array {
+		return Base64Url.decode(base64Url);
 	}
 
 	/**