+++ packages/nameof-transformer/vitest.config.ts
@@ -11,8 +11,7 @@ export default defineConfig({
 		})
 	],
 	test: {
-		include: ["./tests/**/*.ts"],
-		exclude: ["./tests/**/I[A-Z]*.ts"],
+		include: ["./tests/**/*.spec.ts"],
 		globals: true,
 		coverage: {
 			provider: "v8",