+++ packages/core/docs/changelog.md
@@ -3,6 +3,9 @@
 ## v0.0.3
 
 - Added: FactoryInstance class
+- Added: Base64Url class
+- Added: Converter.bytesToBase64Url class
+- Added: Converter.base64UrlToBytes class
 
 ## v0.0.2